Improvements to employee data Cleaned up a bit for newer restclient Moved interface to own file
